Add MTE-2449 TestRail links for TrackingProtection tests (#19345)
* Include TestRail links to the tests

* empty commit

* Fix BasicBrowsing's TR links. Consolidate TrackingProtectionTests

* Move testAdBlocking to TrackingProtectionTest

* Remove TPSidebarBadge.swift

* Update label for "Share Page With"

* Rename TrackingProtectionTest

* testProtectionSidebar has two manual tests

* Remove link to partially automated test

* Rename file TPSettingsTest to TrackingProtectionTest

* Add "Turning TP On/Off via Settings is reflected in Shield menu" test

* Note smoke test

* Add iPad specific step
