+++ focus-ios/focus-ios-tests/FullFunctionalTests.xctestplan
@@ -33,8 +33,6 @@
         "SettingAppearanceTest\/testCheckSetting()",
         "SettingAppearanceTest\/testSafariIntegration()",
         "ShortcutsTest",
-        "TrackingProtectionMenu\/testActiveProtectionSidebar()",
-        "TrackingProtectionMenu\/testProtectionSidebar()",
         "URLValidationTest",
         "WebsiteAccessTests\/testVisitWebsite()"
       ],
